[BUGFIX] Dont use initilize actions for granting access
The initialize actions can not be stoppped anymore, which results in an inpropper access validation
thx to Daniel Hoffmann for providing the patch, thx to the Security Team for reporting it.

https://projekte.in2code.de/issues/60523

+++ Classes/Controller/EditController.php
@@ -18,9 +18,11 @@ use In2code\Femanager\Utility\StringUtility;
 use In2code\Femanager\Utility\UserUtility;
 use Psr\Http\Message\ResponseInterface;
 use TYPO3\CMS\Core\Messaging\AbstractMessage;
+use TYPO3\CMS\Core\Messaging\FlashMessage;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Annotation\Validate;
 use TYPO3\CMS\Extbase\Mvc\Exception\UnsupportedRequestTypeException;
+use TYPO3\CMS\Extbase\Http\ForwardResponse;
 
 /**
  * Class EditController
@@ -44,11 +46,6 @@ class EditController extends AbstractFrontendController
 
     public function initializeUpdateAction()
     {
-        $user = UserUtility::getCurrentUser();
-        $userValues = $this->request->getArgument('user');
-        $token = $this->request->getArgument('token');
-
-        $this->testSpoof($user, $userValues['__identity'], $token);
         if ($this->keepPassword()) {
             unset($this->pluginVariables['user']['password']);
             unset($this->pluginVariables['password_repeat']);
@@ -64,7 +61,29 @@ class EditController extends AbstractFrontendController
      */
     public function updateAction(User $user)
     {
-        $this->redirectIfDirtyObject($user);
+        $currentUser = UserUtility::getCurrentUser();
+        $userValues = $this->request->hasArgument('user') ? $this->request->getArgument('user') : null;
+        $token = $this->request->hasArgument('token') ? $this->request->getArgument('token') : null;
+
+        if ($currentUser === null ||
+            empty($userValues['__identity']) ||
+            (int)$userValues['__identity'] === null ||
+            $token === null ||
+            $this->isSpoof($currentUser, (int)$userValues['__identity'], $token)
+        ) {
+            $this->logUtility->log(Log::STATUS_PROFILEUPDATEREFUSEDSECURITY, $user);
+            $this->addFlashMessage(
+                LocalizationUtility::translateByState(Log::STATUS_PROFILEUPDATEREFUSEDSECURITY),
+                '',
+                FlashMessage::ERROR
+            );
+            return new ForwardResponse('edit');
+        }
+
+        $response = $this->redirectIfNoChangesOnObject($user);
+        if ($response !== null) {
+            return $response;
+        }
         $user = FrontendUtility::forceValues(
             $user,
             ConfigurationUtility::getValue('edit./forceValues./beforeAnyConfirmation.', $this->config)
@@ -172,6 +191,23 @@ class EditController extends AbstractFrontendController
      */
     public function deleteAction(User $user)
     {
+        $currentUser = UserUtility::getCurrentUser();
+        $token = $this->request->hasArgument('token') ? $this->request->getArgument('token') : null;
+        $uid = $this->request->hasArgument('user') ? $this->request->getArgument('user') : null;
+        if ($currentUser === null ||
+            $token === null ||
+            $uid === null ||
+            $this->isSpoof($currentUser, (int)$uid, $token)
+        ) {
+            $this->logUtility->log(Log::STATUS_PROFILEUPDATEREFUSEDSECURITY, $user);
+            $this->addFlashMessage(
+                LocalizationUtility::translateByState(Log::STATUS_PROFILEUPDATEREFUSEDSECURITY),
+                '',
+                FlashMessage::ERROR
+            );
+            return new ForwardResponse('edit');
+        }
+
         $this->eventDispatcher->dispatch(new DeleteUserEvent($user));
         $this->logUtility->log(Log::STATUS_PROFILEDELETE, $user);
         $this->addFlashMessage(LocalizationUtility::translateByState(Log::STATUS_PROFILEDELETE));
@@ -198,16 +234,14 @@ class EditController extends AbstractFrontendController
 
     /**
      * Check: If there are no changes, simple redirect back
-     *
-     * @param User $user
-     * @throws UnsupportedRequestTypeException
      */
-    protected function redirectIfDirtyObject(User $user)
+    protected function redirectIfNoChangesOnObject(User $user)
     {
         if (!ObjectUtility::isDirtyObject($user)) {
-            $this->addFlashMessage(LocalizationUtility::translate('noChanges'), '', AbstractMessage::NOTICE);
-            $this->redirect('edit');
+            $this->addFlashMessage(LocalizationUtility::translate('noChanges'), '', FlashMessage::NOTICE);
+            return $this->redirect('edit');
         }
+        return null;
     }
 
     /**
